geoguess_env/pyproject.toml
[project]
name = "geoguess-env"
version = "0.1.0"
description = "OpenEnv-compatible GeoGuessr RL environment with LLM agents and TRL GRPO training"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "httpx>=0.27",
    "numpy>=1.26",
    "pydantic>=2.7",
    "python-dotenv>=1.0",
    "openenv-core>=0.2.1",
    "reverse_geocoder>=1.5",
    "astral>=3.2",
]

[project.optional-dependencies]
agents = [
    "httpx>=0.27",
]
training = [
    "trl>=0.9",
    "transformers>=4.45",
    "torch>=2.3",
    "datasets>=2.20",
    "accelerate>=0.30",
    "vllm>=0.4",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["geoguess*", "agents*", "client*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
